pyproject.toml
[tool.poetry]
name = "clappy"
version = "0.1.0"
description = "Deity"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
dicey = "^2.0.7"
discord-py = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
flake8 = "^5.0.4"
black = "^22.10.0"
neovim = "^0.3.1"
pynvim = "^0.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
